v5.2 Intefaces
