+++ docs/allclasses-index.html
@@ -2,10 +2,10 @@
 <!-- NewPage -->
 <html lang="de">
 <head>
-<!-- Generated by javadoc (11.0.4) on Fri Mar 06 18:26:13 CET 2020 -->
+<!-- Generated by javadoc (11.0.4) on Tue Mar 10 18:37:45 CET 2020 -->
 <title>All Classes (JavaBotBlockAPI 5.1.12 API)</title>
 <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
-<meta name="dc.created" content="2020-03-06">
+<meta name="dc.created" content="2020-03-10">
 <link rel="stylesheet" type="text/css" href="stylesheet.css" title="Style">
 <link rel="stylesheet" type="text/css" href="jquery/jquery-ui.css" title="Style">
 <script type="text/javascript" src="script.js"></script>
@@ -28,8 +28,8 @@
     catch(err) {
     }
 //-->
-var data = {"i0":2,"i1":2,"i2":32,"i3":2,"i4":2,"i5":4,"i6":2,"i7":8,"i8":4};
-var tabs = {65535:["t0","All Classes"],2:["t2","Class Summary"],4:["t3","Enum Summary"],8:["t4","Exception Summary"],32:["t6","Annotation Types Summary"]};
+var data = {"i0":2,"i1":2,"i2":32,"i3":1,"i4":2,"i5":2,"i6":4,"i7":2,"i8":8,"i9":4};
+var tabs = {65535:["t0","All Classes"],1:["t1","Interface Summary"],2:["t2","Class Summary"],4:["t3","Enum Summary"],8:["t4","Exception Summary"],32:["t6","Annotation Types Summary"]};
 var altColor = "altColor";
 var rowColor = "rowColor";
 var tableTab = "tableTab";
@@ -106,7 +106,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <ul class="blockList">
 <li class="blockList">
 <table class="typeSummary">
-<caption><span id="t0" class="activeTableTab"><span>All Classes</span><span class="tabEnd">&nbsp;</span></span><span id="t2" class="tableTab"><span><a href="javascript:show(2);">Class Summary</a></span><span class="tabEnd">&nbsp;</span></span><span id="t3" class="tableTab"><span><a href="javascript:show(4);">Enum Summary</a></span><span class="tabEnd">&nbsp;</span></span><span id="t4" class="tableTab"><span><a href="javascript:show(8);">Exception Summary</a></span><span class="tabEnd">&nbsp;</span></span><span id="t6" class="tableTab"><span><a href="javascript:show(32);">Annotation Types Summary</a></span><span class="tabEnd">&nbsp;</span></span></caption>
+<caption><span id="t0" class="activeTableTab"><span>All Classes</span><span class="tabEnd">&nbsp;</span></span><span id="t1" class="tableTab"><span><a href="javascript:show(1);">Interface Summary</a></span><span class="tabEnd">&nbsp;</span></span><span id="t2" class="tableTab"><span><a href="javascript:show(2);">Class Summary</a></span><span class="tabEnd">&nbsp;</span></span><span id="t3" class="tableTab"><span><a href="javascript:show(4);">Enum Summary</a></span><span class="tabEnd">&nbsp;</span></span><span id="t4" class="tableTab"><span><a href="javascript:show(8);">Exception Summary</a></span><span class="tabEnd">&nbsp;</span></span><span id="t6" class="tableTab"><span><a href="javascript:show(32);">Annotation Types Summary</a></span><span class="tabEnd">&nbsp;</span></span></caption>
 <tr>
 <th class="colFirst" scope="col">Class</th>
 <th class="colLast" scope="col">Description</th>
@@ -130,37 +130,41 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 </th>
 </tr>
 <tr id="i3" class="rowColor">
-<td class="colFirst"><a href="org/botblock/javabotblockapi/requests/GetBotAction.html" title="class in org.botblock.javabotblockapi.requests">GetBotAction</a></td>
+<td class="colFirst"><a href="org/botblock/javabotblockapi/requests/GetBotAction.html" title="interface in org.botblock.javabotblockapi.requests">GetBotAction</a></td>
 <th class="colLast" scope="row">
-<div class="block">Class used to perform GET actions on the <a href="https://botblock.org/api/docs#bots" target="_blank"><code>/api/bots/:id</code> endpoint</a>.</div>
+<div class="block">Interface used to perform various GET actions towards the <a href="https://botblock.org/api/docs#bots" target="_blank"><code>/api/bots/:id</code></a> endpoint.</div>
 </th>
 </tr>
 <tr id="i4" class="altColor">
-<td class="colFirst"><a href="org/botblock/javabotblockapi/requests/GetListAction.html" title="class in org.botblock.javabotblockapi.requests">GetListAction</a></td>
+<td class="colFirst"><a href="org/botblock/javabotblockapi/internal/GetBotActionImpl.html" title="class in org.botblock.javabotblockapi.internal">GetBotActionImpl</a></td>
+<th class="colLast" scope="row">&nbsp;</th>
+</tr>
+<tr id="i5" class="rowColor">
+<td class="colFirst"><a href="org/botblock/javabotblockapi/internal/GetListAction.html" title="class in org.botblock.javabotblockapi.internal">GetListAction</a></td>
 <th class="colLast" scope="row">
 <div class="block">Class used to perform GET actions on the <a href="https://botblock.org/api/docs#lists" target="_blank"><code>/api/lists</code></a>
  and <a href="https://botblock.org/api/docs#list" target="_blank"><code>/api/lists/:id</code></a> endpoints.</div>
 </th>
 </tr>
-<tr id="i5" class="rowColor">
-<td class="colFirst"><a href="org/botblock/javabotblockapi/requests/GetListAction.ApiField.html" title="enum in org.botblock.javabotblockapi.requests">GetListAction.ApiField</a></td>
+<tr id="i6" class="altColor">
+<td class="colFirst"><a href="org/botblock/javabotblockapi/internal/GetListAction.ApiField.html" title="enum in org.botblock.javabotblockapi.internal">GetListAction.ApiField</a></td>
 <th class="colLast" scope="row">
 <div class="block">Enum containing the different API fields a bot list may have.</div>
 </th>
 </tr>
-<tr id="i6" class="altColor">
-<td class="colFirst"><a href="org/botblock/javabotblockapi/requests/PostAction.html" title="class in org.botblock.javabotblockapi.requests">PostAction</a></td>
+<tr id="i7" class="rowColor">
+<td class="colFirst"><a href="org/botblock/javabotblockapi/internal/PostAction.html" title="class in org.botblock.javabotblockapi.internal">PostAction</a></td>
 <th class="colLast" scope="row">
 <div class="block">Class used to perform POST requests towards the <a href="https://botblock.org/api/docs#count" target="_blank">/api/count</a> endpoint.</div>
 </th>
 </tr>
-<tr id="i7" class="rowColor">
+<tr id="i8" class="altColor">
 <td class="colFirst"><a href="org/botblock/javabotblockapi/exceptions/RatelimitedException.html" title="class in org.botblock.javabotblockapi.exceptions">RatelimitedException</a></td>
 <th class="colLast" scope="row">
 <div class="block">Class used to indicate when the bot gets Ratelimited by the BotBlock API.</div>
 </th>
 </tr>
-<tr id="i8" class="altColor">
+<tr id="i9" class="rowColor">
 <td class="colFirst"><a href="org/botblock/javabotblockapi/Site.html" title="enum in org.botblock.javabotblockapi">Site</a></td>
 <th class="colLast" scope="row">
 <div class="block">Enum class containing all sites currently supported by BotBlock.org.</div>
