+++ docs/allclasses.html
@@ -2,10 +2,10 @@
 <!-- NewPage -->
 <html lang="de">
 <head>
-<!-- Generated by javadoc (11.0.4) on Fri Mar 06 18:26:13 CET 2020 -->
+<!-- Generated by javadoc (11.0.4) on Tue Mar 10 18:37:45 CET 2020 -->
 <title>All Classes (JavaBotBlockAPI 5.1.12 API)</title>
 <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
-<meta name="dc.created" content="2020-03-06">
+<meta name="dc.created" content="2020-03-10">
 <link rel="stylesheet" type="text/css" href="stylesheet.css" title="Style">
 <link rel="stylesheet" type="text/css" href="jquery/jquery-ui.css" title="Style">
 <script type="text/javascript" src="script.js"></script>
@@ -25,10 +25,11 @@
 <li><a href="org/botblock/javabotblockapi/BotBlockAPI.html" title="class in org.botblock.javabotblockapi">BotBlockAPI</a></li>
 <li><a href="org/botblock/javabotblockapi/BotBlockAPI.Builder.html" title="class in org.botblock.javabotblockapi">BotBlockAPI.Builder</a></li>
 <li><a href="org/botblock/javabotblockapi/annotations/DeprecatedSince.html" title="annotation in org.botblock.javabotblockapi.annotations">DeprecatedSince</a></li>
-<li><a href="org/botblock/javabotblockapi/requests/GetBotAction.html" title="class in org.botblock.javabotblockapi.requests">GetBotAction</a></li>
-<li><a href="org/botblock/javabotblockapi/requests/GetListAction.html" title="class in org.botblock.javabotblockapi.requests">GetListAction</a></li>
-<li><a href="org/botblock/javabotblockapi/requests/GetListAction.ApiField.html" title="enum in org.botblock.javabotblockapi.requests">GetListAction.ApiField</a></li>
-<li><a href="org/botblock/javabotblockapi/requests/PostAction.html" title="class in org.botblock.javabotblockapi.requests">PostAction</a></li>
+<li><a href="org/botblock/javabotblockapi/requests/GetBotAction.html" title="interface in org.botblock.javabotblockapi.requests"><span class="interfaceName">GetBotAction</span></a></li>
+<li><a href="org/botblock/javabotblockapi/internal/GetBotActionImpl.html" title="class in org.botblock.javabotblockapi.internal">GetBotActionImpl</a></li>
+<li><a href="org/botblock/javabotblockapi/internal/GetListAction.html" title="class in org.botblock.javabotblockapi.internal">GetListAction</a></li>
+<li><a href="org/botblock/javabotblockapi/internal/GetListAction.ApiField.html" title="enum in org.botblock.javabotblockapi.internal">GetListAction.ApiField</a></li>
+<li><a href="org/botblock/javabotblockapi/internal/PostAction.html" title="class in org.botblock.javabotblockapi.internal">PostAction</a></li>
 <li><a href="org/botblock/javabotblockapi/exceptions/RatelimitedException.html" title="class in org.botblock.javabotblockapi.exceptions">RatelimitedException</a></li>
 <li><a href="org/botblock/javabotblockapi/Site.html" title="enum in org.botblock.javabotblockapi">Site</a></li>
 </ul>
