+++ docs/member-search-index.js
@@ -1,1 +1,1 @@
-memberSearchIndex = [{"p":"org.botblock.javabotblockapi","c":"BotBlockAPI.Builder","l":"addAuthToken(Site, String)","url":"addAuthToken(org.botblock.javabotblockapi.Site,java.lang.String)"},{"p":"org.botblock.javabotblockapi","c":"BotBlockAPI.Builder","l":"addAuthToken(String, String)","url":"addAuthToken(java.lang.String,java.lang.String)"},{"p":"org.botblock.javabotblockapi.requests","c":"GetListAction.ApiField","l":"API_ALL"},{"p":"org.botblock.javabotblockapi.requests","c":"GetListAction.ApiField","l":"API_DOCS"},{"p":"org.botblock.javabotblockapi.requests","c":"GetListAction.ApiField","l":"API_FIELD"},{"p":"org.botblock.javabotblockapi.requests","c":"GetListAction.ApiField","l":"API_GET"},{"p":"org.botblock.javabotblockapi.requests","c":"GetListAction.ApiField","l":"API_POST"},{"p":"org.botblock.javabotblockapi.requests","c":"GetListAction.ApiField","l":"API_SHARD_COUNT"},{"p":"org.botblock.javabotblockapi.requests","c":"GetListAction.ApiField","l":"API_SHARD_ID"},{"p":"org.botblock.javabotblockapi.requests","c":"GetListAction.ApiField","l":"API_SHARDS"},{"p":"org.botblock.javabotblockapi","c":"Site","l":"ARCANE_BOTCENTER_XYZ"},{"p":"org.botblock.javabotblockapi","c":"BotBlockAPI","l":"BotBlockAPI(Map<String, String>, Integer)","url":"%3Cinit%3E(java.util.Map,java.lang.Integer)"},{"p":"org.botblock.javabotblockapi","c":"BotBlockAPI","l":"BotBlockAPI(Map<String, String>)","url":"%3Cinit%3E(java.util.Map)"},{"p":"org.botblock.javabotblockapi","c":"Site","l":"BOTLIST_SPACE"},{"p":"org.botblock.javabotblockapi","c":"Site","l":"BOTS_ONDISCORD_XYZ"},{"p":"org.botblock.javabotblockapi","c":"Site","l":"BOTSFORDISCORD_COM"},{"p":"org.botblock.javabotblockapi","c":"BotBlockAPI.Builder","l":"build()"},{"p":"org.botblock.javabotblockapi","c":"BotBlockAPI.Builder","l":"Builder()","url":"%3Cinit%3E()"},{"p":"org.botblock.javabotblockapi","c":"Site","l":"CLOUD_BOTLIST_XYZ"},{"p":"org.botblock.javabotblockapi","c":"Site","l":"CLOUDLIST_XYZ"},{"p":"org.botblock.javabotblockapi","c":"Site","l":"DBLISTA_PL"},{"p":"org.botblock.javabotblockapi.requests","c":"PostAction","l":"disableAutoPost()"},{"p":"org.botblock.javabotblockapi","c":"Site","l":"DISCORD_BOATS"},{"p":"org.botblock.javabotblockapi","c":"Site","l":"DISCORD_BOTS_GG"},{"p":"org.botblock.javabotblockapi","c":"Site","l":"DISCORDAPPS_DEV"},{"p":"org.botblock.javabotblockapi","c":"Site","l":"DISCORDBOT_WORLD"},{"p":"org.botblock.javabotblockapi","c":"Site","l":"DISCORDBOTLIST_COM"},{"p":"org.botblock.javabotblockapi","c":"Site","l":"DISCORDBOTREVIEWS_XYZ"},{"p":"org.botblock.javabotblockapi","c":"Site","l":"DISCORDBOTS_FUN"},{"p":"org.botblock.javabotblockapi","c":"Site","l":"DISCORDBOTS_ORG"},{"p":"org.botblock.javabotblockapi","c":"Site","l":"DISCORDEXTREMELIST_XYZ"},{"p":"org.botblock.javabotblockapi","c":"Site","l":"DISCORDSBESTBOTS_XYZ"},{"p":"org.botblock.javabotblockapi","c":"Site","l":"DIVINEDISCORDBOTS_COM"},{"p":"org.botblock.javabotblockapi.requests","c":"PostAction","l":"enableAutoPost(JDA, BotBlockAPI)","url":"enableAutoPost(net.dv8tion.jda.api.JDA,org.botblock.javabotblockapi.BotBlockAPI)"},{"p":"org.botblock.javabotblockapi.requests","c":"PostAction","l":"enableAutoPost(Long, int, BotBlockAPI)","url":"enableAutoPost(java.lang.Long,int,org.botblock.javabotblockapi.BotBlockAPI)"},{"p":"org.botblock.javabotblockapi.requests","c":"PostAction","l":"enableAutoPost(ShardManager, BotBlockAPI)","url":"enableAutoPost(net.dv8tion.jda.api.sharding.ShardManager,org.botblock.javabotblockapi.BotBlockAPI)"},{"p":"org.botblock.javabotblockapi.requests","c":"PostAction","l":"enableAutoPost(String, int, BotBlockAPI)","url":"enableAutoPost(java.lang.String,int,org.botblock.javabotblockapi.BotBlockAPI)"},{"p":"org.botblock.javabotblockapi.requests","c":"GetListAction","l":"getApiField(String, Site, GetListAction.ApiField)","url":"getApiField(java.lang.String,org.botblock.javabotblockapi.Site,org.botblock.javabotblockapi.requests.GetListAction.ApiField)"},{"p":"org.botblock.javabotblockapi.requests","c":"GetListAction","l":"getApiField(String, String, GetListAction.ApiField)","url":"getApiField(java.lang.String,java.lang.String,org.botblock.javabotblockapi.requests.GetListAction.ApiField)"},{"p":"org.botblock.javabotblockapi.requests","c":"GetBotAction","l":"GetBotAction()","url":"%3Cinit%3E()"},{"p":"org.botblock.javabotblockapi.requests","c":"GetBotAction","l":"GetBotAction(boolean)","url":"%3Cinit%3E(boolean)"},{"p":"org.botblock.javabotblockapi.exceptions","c":"RatelimitedException","l":"getBotId()"},{"p":"org.botblock.javabotblockapi.requests","c":"GetBotAction","l":"getBotInfo(Long)","url":"getBotInfo(java.lang.Long)"},{"p":"org.botblock.javabotblockapi.requests","c":"GetBotAction","l":"getBotInfo(String)","url":"getBotInfo(java.lang.String)"},{"p":"org.botblock.javabotblockapi.requests","c":"GetBotAction","l":"getBotListInfo(Long, Site)","url":"getBotListInfo(java.lang.Long,org.botblock.javabotblockapi.Site)"},{"p":"org.botblock.javabotblockapi.requests","c":"GetBotAction","l":"getBotListInfo(Long, String)","url":"getBotListInfo(java.lang.Long,java.lang.String)"},{"p":"org.botblock.javabotblockapi.requests","c":"GetBotAction","l":"getBotListInfo(Long)","url":"getBotListInfo(java.lang.Long)"},{"p":"org.botblock.javabotblockapi.requests","c":"GetBotAction","l":"getBotListInfo(String, Site)","url":"getBotListInfo(java.lang.String,org.botblock.javabotblockapi.Site)"},{"p":"org.botblock.javabotblockapi.requests","c":"GetBotAction","l":"getBotListInfo(String, String)","url":"getBotListInfo(java.lang.String,java.lang.String)"},{"p":"org.botblock.javabotblockapi.requests","c":"GetBotAction","l":"getBotListInfo(String)","url":"getBotListInfo(java.lang.String)"},{"p":"org.botblock.javabotblockapi.requests","c":"GetListAction","l":"getBotWidgetUrl(String, Site)","url":"getBotWidgetUrl(java.lang.String,org.botblock.javabotblockapi.Site)"},{"p":"org.botblock.javabotblockapi.requests","c":"GetListAction","l":"getBotWidgetUrl(String, String)","url":"getBotWidgetUrl(java.lang.String,java.lang.String)"},{"p":"org.botblock.javabotblockapi.exceptions","c":"RatelimitedException","l":"getDelay()"},{"p":"org.botblock.javabotblockapi.requests","c":"GetListAction","l":"getDescription(String, Site)","url":"getDescription(java.lang.String,org.botblock.javabotblockapi.Site)"},{"p":"org.botblock.javabotblockapi.requests","c":"GetListAction","l":"getDescription(String, String)","url":"getDescription(java.lang.String,java.lang.String)"},{"p":"org.botblock.javabotblockapi.requests","c":"GetListAction","l":"getDiscordInvite(String, Site)","url":"getDiscordInvite(java.lang.String,org.botblock.javabotblockapi.Site)"},{"p":"org.botblock.javabotblockapi.requests","c":"GetListAction","l":"getDiscordInvite(String, String)","url":"getDiscordInvite(java.lang.String,java.lang.String)"},{"p":"org.botblock.javabotblockapi.requests","c":"GetBotAction","l":"getDiscriminator(Long)","url":"getDiscriminator(java.lang.Long)"},{"p":"org.botblock.javabotblockapi.requests","c":"GetBotAction","l":"getDiscriminator(String)","url":"getDiscriminator(java.lang.String)"},{"p":"org.botblock.javabotblockapi.requests","c":"GetListAction","l":"getFeatures(String, Site)","url":"getFeatures(java.lang.String,org.botblock.javabotblockapi.Site)"},{"p":"org.botblock.javabotblockapi.requests","c":"GetListAction","l":"getFeatures(String, String)","url":"getFeatures(java.lang.String,java.lang.String)"},{"p":"org.botblock.javabotblockapi.requests","c":"GetListAction","l":"getFilteredLists(String)","url":"getFilteredLists(java.lang.String)"},{"p":"org.botblock.javabotblockapi.requests","c":"GetBotAction","l":"getGitHub(Long)","url":"getGitHub(java.lang.Long)"},{"p":"org.botblock.javabotblockapi.requests","c":"GetBotAction","l":"getGitHub(String)","url":"getGitHub(java.lang.String)"},{"p":"org.botblock.javabotblockapi.requests","c":"GetListAction","l":"getIcon(String, Site)","url":"getIcon(java.lang.String,org.botblock.javabotblockapi.Site)"},{"p":"org.botblock.javabotblockapi.requests","c":"GetListAction","l":"getIcon(String, String)","url":"getIcon(java.lang.String,java.lang.String)"},{"p":"org.botblock.javabotblockapi.requests","c":"GetListAction","l":"getId(String, Site)","url":"getId(java.lang.String,org.botblock.javabotblockapi.Site)"},{"p":"org.botblock.javabotblockapi.requests","c":"GetListAction","l":"getId(String, String)","url":"getId(java.lang.String,java.lang.String)"},{"p":"org.botblock.javabotblockapi.requests","c":"GetBotAction","l":"getInvite(Long)","url":"getInvite(java.lang.Long)"},{"p":"org.botblock.javabotblockapi.requests","c":"GetBotAction","l":"getInvite(String)","url":"getInvite(java.lang.String)"},{"p":"org.botblock.javabotblockapi.exceptions","c":"RatelimitedException","l":"getIp()"},{"p":"org.botblock.javabotblockapi.requests","c":"GetListAction","l":"getLanguage(String, Site)","url":"getLanguage(java.lang.String,org.botblock.javabotblockapi.Site)"},{"p":"org.botblock.javabotblockapi.requests","c":"GetListAction","l":"getLanguage(String, String)","url":"getLanguage(java.lang.String,java.lang.String)"},{"p":"org.botblock.javabotblockapi.requests","c":"GetBotAction","l":"getLibrary(Long)","url":"getLibrary(java.lang.Long)"},{"p":"org.botblock.javabotblockapi.requests","c":"GetBotAction","l":"getLibrary(String)","url":"getLibrary(java.lang.String)"},{"p":"org.botblock.javabotblockapi.requests","c":"GetListAction","l":"getList(String, Site)","url":"getList(java.lang.String,org.botblock.javabotblockapi.Site)"},{"p":"org.botblock.javabotblockapi.requests","c":"GetListAction","l":"getList(String, String)","url":"getList(java.lang.String,java.lang.String)"},{"p":"org.botblock.javabotblockapi.requests","c":"GetListAction","l":"GetListAction()","url":"%3Cinit%3E()"},{"p":"org.botblock.javabotblockapi.requests","c":"GetListAction","l":"GetListAction(boolean)","url":"%3Cinit%3E(boolean)"},{"p":"org.botblock.javabotblockapi.requests","c":"GetListAction","l":"getLists(String)","url":"getLists(java.lang.String)"},{"p":"org.botblock.javabotblockapi.exceptions","c":"RatelimitedException","l":"getMessage()"},{"p":"org.botblock.javabotblockapi.requests","c":"GetBotAction","l":"getName(Long)","url":"getName(java.lang.Long)"},{"p":"org.botblock.javabotblockapi.requests","c":"GetListAction","l":"getName(String, Site)","url":"getName(java.lang.String,org.botblock.javabotblockapi.Site)"},{"p":"org.botblock.javabotblockapi.requests","c":"GetListAction","l":"getName(String, String)","url":"getName(java.lang.String,java.lang.String)"},{"p":"org.botblock.javabotblockapi.requests","c":"GetBotAction","l":"getName(String)","url":"getName(java.lang.String)"},{"p":"org.botblock.javabotblockapi.requests","c":"GetBotAction","l":"getOwners(Long)","url":"getOwners(java.lang.Long)"},{"p":"org.botblock.javabotblockapi.requests","c":"GetListAction","l":"getOwners(String, Site)","url":"getOwners(java.lang.String,org.botblock.javabotblockapi.Site)"},{"p":"org.botblock.javabotblockapi.requests","c":"GetListAction","l":"getOwners(String, String)","url":"getOwners(java.lang.String,java.lang.String)"},{"p":"org.botblock.javabotblockapi.requests","c":"GetBotAction","l":"getOwners(String)","url":"getOwners(java.lang.String)"},{"p":"org.botblock.javabotblockapi.requests","c":"GetBotAction","l":"getPrefix(Long)","url":"getPrefix(java.lang.Long)"},{"p":"org.botblock.javabotblockapi.requests","c":"GetBotAction","l":"getPrefix(String)","url":"getPrefix(java.lang.String)"},{"p":"org.botblock.javabotblockapi.exceptions","c":"RatelimitedException","l":"getRoute()"},{"p":"org.botblock.javabotblockapi.requests","c":"GetBotAction","l":"getServerCount(Long)","url":"getServerCount(java.lang.Long)"},{"p":"org.botblock.javabotblockapi.requests","c":"GetBotAction","l":"getServerCount(String)","url":"getServerCount(java.lang.String)"},{"p":"org.botblock.javabotblockapi","c":"Site","l":"getSite()"},{"p":"org.botblock.javabotblockapi.requests","c":"GetBotAction","l":"getSupportLink(Long)","url":"getSupportLink(java.lang.Long)"},{"p":"org.botblock.javabotblockapi.requests","c":"GetBotAction","l":"getSupportLink(String)","url":"getSupportLink(java.lang.String)"},{"p":"org.botblock.javabotblockapi.requests","c":"GetListAction","l":"getTimeAdded(String, Site)","url":"getTimeAdded(java.lang.String,org.botblock.javabotblockapi.Site)"},{"p":"org.botblock.javabotblockapi.requests","c":"GetListAction","l":"getTimeAdded(String, String)","url":"getTimeAdded(java.lang.String,java.lang.String)"},{"p":"org.botblock.javabotblockapi","c":"BotBlockAPI","l":"getTokens()"},{"p":"org.botblock.javabotblockapi","c":"BotBlockAPI","l":"getUpdateDelay()"},{"p":"org.botblock.javabotblockapi.requests","c":"GetListAction","l":"getUrl(String, Site)","url":"getUrl(java.lang.String,org.botblock.javabotblockapi.Site)"},{"p":"org.botblock.javabotblockapi.requests","c":"GetListAction","l":"getUrl(String, String)","url":"getUrl(java.lang.String,java.lang.String)"},{"p":"org.botblock.javabotblockapi.requests","c":"GetBotAction","l":"getWebsite(Long)","url":"getWebsite(java.lang.Long)"},{"p":"org.botblock.javabotblockapi.requests","c":"GetBotAction","l":"getWebsite(String)","url":"getWebsite(java.lang.String)"},{"p":"org.botblock.javabotblockapi","c":"Site","l":"GLENNBOTLIST_XYZ"},{"p":"org.botblock.javabotblockapi.requests","c":"GetListAction","l":"isDefunct(String, Site)","url":"isDefunct(java.lang.String,org.botblock.javabotblockapi.Site)"},{"p":"org.botblock.javabotblockapi.requests","c":"GetListAction","l":"isDefunct(String, String)","url":"isDefunct(java.lang.String,java.lang.String)"},{"p":"org.botblock.javabotblockapi.requests","c":"GetListAction","l":"isDiscordOnly(String, Site)","url":"isDiscordOnly(java.lang.String,org.botblock.javabotblockapi.Site)"},{"p":"org.botblock.javabotblockapi.requests","c":"GetListAction","l":"isDiscordOnly(String, String)","url":"isDiscordOnly(java.lang.String,java.lang.String)"},{"p":"org.botblock.javabotblockapi","c":"Site","l":"LBOTS_ORG"},{"p":"org.botblock.javabotblockapi","c":"Site","l":"MYTHICALBOTS_XYZ"},{"p":"org.botblock.javabotblockapi.requests","c":"PostAction","l":"PostAction()","url":"%3Cinit%3E()"},{"p":"org.botblock.javabotblockapi.requests","c":"PostAction","l":"postGuilds(JDA, BotBlockAPI)","url":"postGuilds(net.dv8tion.jda.api.JDA,org.botblock.javabotblockapi.BotBlockAPI)"},{"p":"org.botblock.javabotblockapi.requests","c":"PostAction","l":"postGuilds(Long, Integer, BotBlockAPI)","url":"postGuilds(java.lang.Long,java.lang.Integer,org.botblock.javabotblockapi.BotBlockAPI)"},{"p":"org.botblock.javabotblockapi.requests","c":"PostAction","l":"postGuilds(ShardManager, BotBlockAPI)","url":"postGuilds(net.dv8tion.jda.api.sharding.ShardManager,org.botblock.javabotblockapi.BotBlockAPI)"},{"p":"org.botblock.javabotblockapi.requests","c":"PostAction","l":"postGuilds(String, Integer, BotBlockAPI)","url":"postGuilds(java.lang.String,java.lang.Integer,org.botblock.javabotblockapi.BotBlockAPI)"},{"p":"org.botblock.javabotblockapi.exceptions","c":"RatelimitedException","l":"RatelimitedException(String)","url":"%3Cinit%3E(java.lang.String)"},{"p":"org.botblock.javabotblockapi.annotations","c":"DeprecatedSince","l":"replacements()"},{"p":"org.botblock.javabotblockapi","c":"BotBlockAPI.Builder","l":"setAuthTokens(Map<String, String>)","url":"setAuthTokens(java.util.Map)"},{"p":"org.botblock.javabotblockapi","c":"BotBlockAPI.Builder","l":"setUpdateDelay(Integer)","url":"setUpdateDelay(java.lang.Integer)"},{"p":"org.botblock.javabotblockapi","c":"Site","l":"SPACE_BOT_LIST_ORG"},{"p":"org.botblock.javabotblockapi","c":"Site","l":"TOP_GG"},{"p":"org.botblock.javabotblockapi.exceptions","c":"RatelimitedException","l":"toString()"},{"p":"org.botblock.javabotblockapi","c":"Site","l":"valueOf(String)","url":"valueOf(java.lang.String)"},{"p":"org.botblock.javabotblockapi.requests","c":"GetListAction.ApiField","l":"valueOf(String)","url":"valueOf(java.lang.String)"},{"p":"org.botblock.javabotblockapi","c":"Site","l":"values()"},{"p":"org.botblock.javabotblockapi.requests","c":"GetListAction.ApiField","l":"values()"},{"p":"org.botblock.javabotblockapi.annotations","c":"DeprecatedSince","l":"version()"},{"p":"org.botblock.javabotblockapi","c":"Site","l":"WONDERBOTLIST_COM"},{"p":"org.botblock.javabotblockapi","c":"Site","l":"YABL_XYZ"}]
+memberSearchIndex = [{"p":"org.botblock.javabotblockapi","c":"BotBlockAPI.Builder","l":"addAuthToken(Site, String)","url":"addAuthToken(org.botblock.javabotblockapi.Site,java.lang.String)"},{"p":"org.botblock.javabotblockapi","c":"BotBlockAPI.Builder","l":"addAuthToken(String, String)","url":"addAuthToken(java.lang.String,java.lang.String)"},{"p":"org.botblock.javabotblockapi.internal","c":"GetListAction.ApiField","l":"API_ALL"},{"p":"org.botblock.javabotblockapi.internal","c":"GetListAction.ApiField","l":"API_DOCS"},{"p":"org.botblock.javabotblockapi.internal","c":"GetListAction.ApiField","l":"API_FIELD"},{"p":"org.botblock.javabotblockapi.internal","c":"GetListAction.ApiField","l":"API_GET"},{"p":"org.botblock.javabotblockapi.internal","c":"GetListAction.ApiField","l":"API_POST"},{"p":"org.botblock.javabotblockapi.internal","c":"GetListAction.ApiField","l":"API_SHARD_COUNT"},{"p":"org.botblock.javabotblockapi.internal","c":"GetListAction.ApiField","l":"API_SHARD_ID"},{"p":"org.botblock.javabotblockapi.internal","c":"GetListAction.ApiField","l":"API_SHARDS"},{"p":"org.botblock.javabotblockapi","c":"Site","l":"ARCANE_BOTCENTER_XYZ"},{"p":"org.botblock.javabotblockapi","c":"BotBlockAPI","l":"BotBlockAPI(Map<String, String>, Integer)","url":"%3Cinit%3E(java.util.Map,java.lang.Integer)"},{"p":"org.botblock.javabotblockapi","c":"BotBlockAPI","l":"BotBlockAPI(Map<String, String>)","url":"%3Cinit%3E(java.util.Map)"},{"p":"org.botblock.javabotblockapi","c":"Site","l":"BOTLIST_SPACE"},{"p":"org.botblock.javabotblockapi","c":"Site","l":"BOTS_ONDISCORD_XYZ"},{"p":"org.botblock.javabotblockapi","c":"Site","l":"BOTSFORDISCORD_COM"},{"p":"org.botblock.javabotblockapi","c":"BotBlockAPI.Builder","l":"build()"},{"p":"org.botblock.javabotblockapi","c":"BotBlockAPI.Builder","l":"Builder()","url":"%3Cinit%3E()"},{"p":"org.botblock.javabotblockapi","c":"Site","l":"CLOUD_BOTLIST_XYZ"},{"p":"org.botblock.javabotblockapi","c":"Site","l":"CLOUDLIST_XYZ"},{"p":"org.botblock.javabotblockapi","c":"Site","l":"DBLISTA_PL"},{"p":"org.botblock.javabotblockapi.internal","c":"PostAction","l":"disableAutoPost()"},{"p":"org.botblock.javabotblockapi","c":"Site","l":"DISCORD_BOATS"},{"p":"org.botblock.javabotblockapi","c":"Site","l":"DISCORD_BOTS_GG"},{"p":"org.botblock.javabotblockapi","c":"Site","l":"DISCORDAPPS_DEV"},{"p":"org.botblock.javabotblockapi","c":"Site","l":"DISCORDBOT_WORLD"},{"p":"org.botblock.javabotblockapi","c":"Site","l":"DISCORDBOTLIST_COM"},{"p":"org.botblock.javabotblockapi","c":"Site","l":"DISCORDBOTREVIEWS_XYZ"},{"p":"org.botblock.javabotblockapi","c":"Site","l":"DISCORDBOTS_FUN"},{"p":"org.botblock.javabotblockapi","c":"Site","l":"DISCORDBOTS_ORG"},{"p":"org.botblock.javabotblockapi","c":"Site","l":"DISCORDEXTREMELIST_XYZ"},{"p":"org.botblock.javabotblockapi","c":"Site","l":"DISCORDSBESTBOTS_XYZ"},{"p":"org.botblock.javabotblockapi","c":"Site","l":"DIVINEDISCORDBOTS_COM"},{"p":"org.botblock.javabotblockapi.internal","c":"PostAction","l":"enableAutoPost(JDA, BotBlockAPI)","url":"enableAutoPost(net.dv8tion.jda.api.JDA,org.botblock.javabotblockapi.BotBlockAPI)"},{"p":"org.botblock.javabotblockapi.internal","c":"PostAction","l":"enableAutoPost(Long, int, BotBlockAPI)","url":"enableAutoPost(java.lang.Long,int,org.botblock.javabotblockapi.BotBlockAPI)"},{"p":"org.botblock.javabotblockapi.internal","c":"PostAction","l":"enableAutoPost(ShardManager, BotBlockAPI)","url":"enableAutoPost(net.dv8tion.jda.api.sharding.ShardManager,org.botblock.javabotblockapi.BotBlockAPI)"},{"p":"org.botblock.javabotblockapi.internal","c":"PostAction","l":"enableAutoPost(String, int, BotBlockAPI)","url":"enableAutoPost(java.lang.String,int,org.botblock.javabotblockapi.BotBlockAPI)"},{"p":"org.botblock.javabotblockapi.internal","c":"GetListAction","l":"getApiField(String, Site, GetListAction.ApiField)","url":"getApiField(java.lang.String,org.botblock.javabotblockapi.Site,org.botblock.javabotblockapi.internal.GetListAction.ApiField)"},{"p":"org.botblock.javabotblockapi.internal","c":"GetListAction","l":"getApiField(String, String, GetListAction.ApiField)","url":"getApiField(java.lang.String,java.lang.String,org.botblock.javabotblockapi.internal.GetListAction.ApiField)"},{"p":"org.botblock.javabotblockapi.internal","c":"GetBotActionImpl","l":"GetBotActionImpl()","url":"%3Cinit%3E()"},{"p":"org.botblock.javabotblockapi.internal","c":"GetBotActionImpl","l":"GetBotActionImpl(boolean)","url":"%3Cinit%3E(boolean)"},{"p":"org.botblock.javabotblockapi.exceptions","c":"RatelimitedException","l":"getBotId()"},{"p":"org.botblock.javabotblockapi.internal","c":"GetBotActionImpl","l":"getBotInfo(Long)","url":"getBotInfo(java.lang.Long)"},{"p":"org.botblock.javabotblockapi.requests","c":"GetBotAction","l":"getBotInfo(Long)","url":"getBotInfo(java.lang.Long)"},{"p":"org.botblock.javabotblockapi.internal","c":"GetBotActionImpl","l":"getBotInfo(String)","url":"getBotInfo(java.lang.String)"},{"p":"org.botblock.javabotblockapi.requests","c":"GetBotAction","l":"getBotInfo(String)","url":"getBotInfo(java.lang.String)"},{"p":"org.botblock.javabotblockapi.internal","c":"GetBotActionImpl","l":"getBotListInfo(Long, Site)","url":"getBotListInfo(java.lang.Long,org.botblock.javabotblockapi.Site)"},{"p":"org.botblock.javabotblockapi.requests","c":"GetBotAction","l":"getBotListInfo(Long, Site)","url":"getBotListInfo(java.lang.Long,org.botblock.javabotblockapi.Site)"},{"p":"org.botblock.javabotblockapi.internal","c":"GetBotActionImpl","l":"getBotListInfo(Long, String)","url":"getBotListInfo(java.lang.Long,java.lang.String)"},{"p":"org.botblock.javabotblockapi.requests","c":"GetBotAction","l":"getBotListInfo(Long, String)","url":"getBotListInfo(java.lang.Long,java.lang.String)"},{"p":"org.botblock.javabotblockapi.internal","c":"GetBotActionImpl","l":"getBotListInfo(Long)","url":"getBotListInfo(java.lang.Long)"},{"p":"org.botblock.javabotblockapi.requests","c":"GetBotAction","l":"getBotListInfo(Long)","url":"getBotListInfo(java.lang.Long)"},{"p":"org.botblock.javabotblockapi.internal","c":"GetBotActionImpl","l":"getBotListInfo(String, Site)","url":"getBotListInfo(java.lang.String,org.botblock.javabotblockapi.Site)"},{"p":"org.botblock.javabotblockapi.requests","c":"GetBotAction","l":"getBotListInfo(String, Site)","url":"getBotListInfo(java.lang.String,org.botblock.javabotblockapi.Site)"},{"p":"org.botblock.javabotblockapi.internal","c":"GetBotActionImpl","l":"getBotListInfo(String, String)","url":"getBotListInfo(java.lang.String,java.lang.String)"},{"p":"org.botblock.javabotblockapi.requests","c":"GetBotAction","l":"getBotListInfo(String, String)","url":"getBotListInfo(java.lang.String,java.lang.String)"},{"p":"org.botblock.javabotblockapi.internal","c":"GetBotActionImpl","l":"getBotListInfo(String)","url":"getBotListInfo(java.lang.String)"},{"p":"org.botblock.javabotblockapi.requests","c":"GetBotAction","l":"getBotListInfo(String)","url":"getBotListInfo(java.lang.String)"},{"p":"org.botblock.javabotblockapi.internal","c":"GetListAction","l":"getBotWidgetUrl(String, Site)","url":"getBotWidgetUrl(java.lang.String,org.botblock.javabotblockapi.Site)"},{"p":"org.botblock.javabotblockapi.internal","c":"GetListAction","l":"getBotWidgetUrl(String, String)","url":"getBotWidgetUrl(java.lang.String,java.lang.String)"},{"p":"org.botblock.javabotblockapi.exceptions","c":"RatelimitedException","l":"getDelay()"},{"p":"org.botblock.javabotblockapi.internal","c":"GetListAction","l":"getDescription(String, Site)","url":"getDescription(java.lang.String,org.botblock.javabotblockapi.Site)"},{"p":"org.botblock.javabotblockapi.internal","c":"GetListAction","l":"getDescription(String, String)","url":"getDescription(java.lang.String,java.lang.String)"},{"p":"org.botblock.javabotblockapi.internal","c":"GetListAction","l":"getDiscordInvite(String, Site)","url":"getDiscordInvite(java.lang.String,org.botblock.javabotblockapi.Site)"},{"p":"org.botblock.javabotblockapi.internal","c":"GetListAction","l":"getDiscordInvite(String, String)","url":"getDiscordInvite(java.lang.String,java.lang.String)"},{"p":"org.botblock.javabotblockapi.internal","c":"GetBotActionImpl","l":"getDiscriminator(Long)","url":"getDiscriminator(java.lang.Long)"},{"p":"org.botblock.javabotblockapi.requests","c":"GetBotAction","l":"getDiscriminator(Long)","url":"getDiscriminator(java.lang.Long)"},{"p":"org.botblock.javabotblockapi.internal","c":"GetBotActionImpl","l":"getDiscriminator(String)","url":"getDiscriminator(java.lang.String)"},{"p":"org.botblock.javabotblockapi.requests","c":"GetBotAction","l":"getDiscriminator(String)","url":"getDiscriminator(java.lang.String)"},{"p":"org.botblock.javabotblockapi.internal","c":"GetListAction","l":"getFeatures(String, Site)","url":"getFeatures(java.lang.String,org.botblock.javabotblockapi.Site)"},{"p":"org.botblock.javabotblockapi.internal","c":"GetListAction","l":"getFeatures(String, String)","url":"getFeatures(java.lang.String,java.lang.String)"},{"p":"org.botblock.javabotblockapi.internal","c":"GetListAction","l":"getFilteredLists(String)","url":"getFilteredLists(java.lang.String)"},{"p":"org.botblock.javabotblockapi.internal","c":"GetBotActionImpl","l":"getGitHub(Long)","url":"getGitHub(java.lang.Long)"},{"p":"org.botblock.javabotblockapi.requests","c":"GetBotAction","l":"getGitHub(Long)","url":"getGitHub(java.lang.Long)"},{"p":"org.botblock.javabotblockapi.internal","c":"GetBotActionImpl","l":"getGitHub(String)","url":"getGitHub(java.lang.String)"},{"p":"org.botblock.javabotblockapi.requests","c":"GetBotAction","l":"getGitHub(String)","url":"getGitHub(java.lang.String)"},{"p":"org.botblock.javabotblockapi.internal","c":"GetListAction","l":"getIcon(String, Site)","url":"getIcon(java.lang.String,org.botblock.javabotblockapi.Site)"},{"p":"org.botblock.javabotblockapi.internal","c":"GetListAction","l":"getIcon(String, String)","url":"getIcon(java.lang.String,java.lang.String)"},{"p":"org.botblock.javabotblockapi.internal","c":"GetListAction","l":"getId(String, Site)","url":"getId(java.lang.String,org.botblock.javabotblockapi.Site)"},{"p":"org.botblock.javabotblockapi.internal","c":"GetListAction","l":"getId(String, String)","url":"getId(java.lang.String,java.lang.String)"},{"p":"org.botblock.javabotblockapi.internal","c":"GetBotActionImpl","l":"getInvite(Long)","url":"getInvite(java.lang.Long)"},{"p":"org.botblock.javabotblockapi.requests","c":"GetBotAction","l":"getInvite(Long)","url":"getInvite(java.lang.Long)"},{"p":"org.botblock.javabotblockapi.internal","c":"GetBotActionImpl","l":"getInvite(String)","url":"getInvite(java.lang.String)"},{"p":"org.botblock.javabotblockapi.requests","c":"GetBotAction","l":"getInvite(String)","url":"getInvite(java.lang.String)"},{"p":"org.botblock.javabotblockapi.exceptions","c":"RatelimitedException","l":"getIp()"},{"p":"org.botblock.javabotblockapi.internal","c":"GetListAction","l":"getLanguage(String, Site)","url":"getLanguage(java.lang.String,org.botblock.javabotblockapi.Site)"},{"p":"org.botblock.javabotblockapi.internal","c":"GetListAction","l":"getLanguage(String, String)","url":"getLanguage(java.lang.String,java.lang.String)"},{"p":"org.botblock.javabotblockapi.internal","c":"GetBotActionImpl","l":"getLibrary(Long)","url":"getLibrary(java.lang.Long)"},{"p":"org.botblock.javabotblockapi.requests","c":"GetBotAction","l":"getLibrary(Long)","url":"getLibrary(java.lang.Long)"},{"p":"org.botblock.javabotblockapi.internal","c":"GetBotActionImpl","l":"getLibrary(String)","url":"getLibrary(java.lang.String)"},{"p":"org.botblock.javabotblockapi.requests","c":"GetBotAction","l":"getLibrary(String)","url":"getLibrary(java.lang.String)"},{"p":"org.botblock.javabotblockapi.internal","c":"GetListAction","l":"getList(String, Site)","url":"getList(java.lang.String,org.botblock.javabotblockapi.Site)"},{"p":"org.botblock.javabotblockapi.internal","c":"GetListAction","l":"getList(String, String)","url":"getList(java.lang.String,java.lang.String)"},{"p":"org.botblock.javabotblockapi.internal","c":"GetListAction","l":"GetListAction()","url":"%3Cinit%3E()"},{"p":"org.botblock.javabotblockapi.internal","c":"GetListAction","l":"GetListAction(boolean)","url":"%3Cinit%3E(boolean)"},{"p":"org.botblock.javabotblockapi.internal","c":"GetListAction","l":"getLists(String)","url":"getLists(java.lang.String)"},{"p":"org.botblock.javabotblockapi.exceptions","c":"RatelimitedException","l":"getMessage()"},{"p":"org.botblock.javabotblockapi.internal","c":"GetBotActionImpl","l":"getName(Long)","url":"getName(java.lang.Long)"},{"p":"org.botblock.javabotblockapi.requests","c":"GetBotAction","l":"getName(Long)","url":"getName(java.lang.Long)"},{"p":"org.botblock.javabotblockapi.internal","c":"GetListAction","l":"getName(String, Site)","url":"getName(java.lang.String,org.botblock.javabotblockapi.Site)"},{"p":"org.botblock.javabotblockapi.internal","c":"GetListAction","l":"getName(String, String)","url":"getName(java.lang.String,java.lang.String)"},{"p":"org.botblock.javabotblockapi.internal","c":"GetBotActionImpl","l":"getName(String)","url":"getName(java.lang.String)"},{"p":"org.botblock.javabotblockapi.requests","c":"GetBotAction","l":"getName(String)","url":"getName(java.lang.String)"},{"p":"org.botblock.javabotblockapi.internal","c":"GetBotActionImpl","l":"getOwners(Long)","url":"getOwners(java.lang.Long)"},{"p":"org.botblock.javabotblockapi.requests","c":"GetBotAction","l":"getOwners(Long)","url":"getOwners(java.lang.Long)"},{"p":"org.botblock.javabotblockapi.internal","c":"GetListAction","l":"getOwners(String, Site)","url":"getOwners(java.lang.String,org.botblock.javabotblockapi.Site)"},{"p":"org.botblock.javabotblockapi.internal","c":"GetListAction","l":"getOwners(String, String)","url":"getOwners(java.lang.String,java.lang.String)"},{"p":"org.botblock.javabotblockapi.internal","c":"GetBotActionImpl","l":"getOwners(String)","url":"getOwners(java.lang.String)"},{"p":"org.botblock.javabotblockapi.requests","c":"GetBotAction","l":"getOwners(String)","url":"getOwners(java.lang.String)"},{"p":"org.botblock.javabotblockapi.internal","c":"GetBotActionImpl","l":"getPrefix(Long)","url":"getPrefix(java.lang.Long)"},{"p":"org.botblock.javabotblockapi.requests","c":"GetBotAction","l":"getPrefix(Long)","url":"getPrefix(java.lang.Long)"},{"p":"org.botblock.javabotblockapi.internal","c":"GetBotActionImpl","l":"getPrefix(String)","url":"getPrefix(java.lang.String)"},{"p":"org.botblock.javabotblockapi.requests","c":"GetBotAction","l":"getPrefix(String)","url":"getPrefix(java.lang.String)"},{"p":"org.botblock.javabotblockapi.exceptions","c":"RatelimitedException","l":"getRoute()"},{"p":"org.botblock.javabotblockapi.internal","c":"GetBotActionImpl","l":"getServerCount(Long)","url":"getServerCount(java.lang.Long)"},{"p":"org.botblock.javabotblockapi.requests","c":"GetBotAction","l":"getServerCount(Long)","url":"getServerCount(java.lang.Long)"},{"p":"org.botblock.javabotblockapi.internal","c":"GetBotActionImpl","l":"getServerCount(String)","url":"getServerCount(java.lang.String)"},{"p":"org.botblock.javabotblockapi.requests","c":"GetBotAction","l":"getServerCount(String)","url":"getServerCount(java.lang.String)"},{"p":"org.botblock.javabotblockapi","c":"Site","l":"getSite()"},{"p":"org.botblock.javabotblockapi.internal","c":"GetBotActionImpl","l":"getSupportLink(Long)","url":"getSupportLink(java.lang.Long)"},{"p":"org.botblock.javabotblockapi.requests","c":"GetBotAction","l":"getSupportLink(Long)","url":"getSupportLink(java.lang.Long)"},{"p":"org.botblock.javabotblockapi.internal","c":"GetBotActionImpl","l":"getSupportLink(String)","url":"getSupportLink(java.lang.String)"},{"p":"org.botblock.javabotblockapi.requests","c":"GetBotAction","l":"getSupportLink(String)","url":"getSupportLink(java.lang.String)"},{"p":"org.botblock.javabotblockapi.internal","c":"GetListAction","l":"getTimeAdded(String, Site)","url":"getTimeAdded(java.lang.String,org.botblock.javabotblockapi.Site)"},{"p":"org.botblock.javabotblockapi.internal","c":"GetListAction","l":"getTimeAdded(String, String)","url":"getTimeAdded(java.lang.String,java.lang.String)"},{"p":"org.botblock.javabotblockapi","c":"BotBlockAPI","l":"getTokens()"},{"p":"org.botblock.javabotblockapi","c":"BotBlockAPI","l":"getUpdateDelay()"},{"p":"org.botblock.javabotblockapi.internal","c":"GetListAction","l":"getUrl(String, Site)","url":"getUrl(java.lang.String,org.botblock.javabotblockapi.Site)"},{"p":"org.botblock.javabotblockapi.internal","c":"GetListAction","l":"getUrl(String, String)","url":"getUrl(java.lang.String,java.lang.String)"},{"p":"org.botblock.javabotblockapi.internal","c":"GetBotActionImpl","l":"getWebsite(Long)","url":"getWebsite(java.lang.Long)"},{"p":"org.botblock.javabotblockapi.requests","c":"GetBotAction","l":"getWebsite(Long)","url":"getWebsite(java.lang.Long)"},{"p":"org.botblock.javabotblockapi.internal","c":"GetBotActionImpl","l":"getWebsite(String)","url":"getWebsite(java.lang.String)"},{"p":"org.botblock.javabotblockapi.requests","c":"GetBotAction","l":"getWebsite(String)","url":"getWebsite(java.lang.String)"},{"p":"org.botblock.javabotblockapi","c":"Site","l":"GLENNBOTLIST_XYZ"},{"p":"org.botblock.javabotblockapi.internal","c":"GetListAction","l":"isDefunct(String, Site)","url":"isDefunct(java.lang.String,org.botblock.javabotblockapi.Site)"},{"p":"org.botblock.javabotblockapi.internal","c":"GetListAction","l":"isDefunct(String, String)","url":"isDefunct(java.lang.String,java.lang.String)"},{"p":"org.botblock.javabotblockapi.internal","c":"GetListAction","l":"isDiscordOnly(String, Site)","url":"isDiscordOnly(java.lang.String,org.botblock.javabotblockapi.Site)"},{"p":"org.botblock.javabotblockapi.internal","c":"GetListAction","l":"isDiscordOnly(String, String)","url":"isDiscordOnly(java.lang.String,java.lang.String)"},{"p":"org.botblock.javabotblockapi","c":"Site","l":"LBOTS_ORG"},{"p":"org.botblock.javabotblockapi","c":"Site","l":"MYTHICALBOTS_XYZ"},{"p":"org.botblock.javabotblockapi.internal","c":"PostAction","l":"PostAction()","url":"%3Cinit%3E()"},{"p":"org.botblock.javabotblockapi.internal","c":"PostAction","l":"postGuilds(JDA, BotBlockAPI)","url":"postGuilds(net.dv8tion.jda.api.JDA,org.botblock.javabotblockapi.BotBlockAPI)"},{"p":"org.botblock.javabotblockapi.internal","c":"PostAction","l":"postGuilds(Long, Integer, BotBlockAPI)","url":"postGuilds(java.lang.Long,java.lang.Integer,org.botblock.javabotblockapi.BotBlockAPI)"},{"p":"org.botblock.javabotblockapi.internal","c":"PostAction","l":"postGuilds(ShardManager, BotBlockAPI)","url":"postGuilds(net.dv8tion.jda.api.sharding.ShardManager,org.botblock.javabotblockapi.BotBlockAPI)"},{"p":"org.botblock.javabotblockapi.internal","c":"PostAction","l":"postGuilds(String, Integer, BotBlockAPI)","url":"postGuilds(java.lang.String,java.lang.Integer,org.botblock.javabotblockapi.BotBlockAPI)"},{"p":"org.botblock.javabotblockapi.exceptions","c":"RatelimitedException","l":"RatelimitedException(String)","url":"%3Cinit%3E(java.lang.String)"},{"p":"org.botblock.javabotblockapi.annotations","c":"DeprecatedSince","l":"replacements()"},{"p":"org.botblock.javabotblockapi","c":"BotBlockAPI.Builder","l":"setAuthTokens(Map<String, String>)","url":"setAuthTokens(java.util.Map)"},{"p":"org.botblock.javabotblockapi","c":"BotBlockAPI.Builder","l":"setUpdateDelay(Integer)","url":"setUpdateDelay(java.lang.Integer)"},{"p":"org.botblock.javabotblockapi","c":"Site","l":"SPACE_BOT_LIST_ORG"},{"p":"org.botblock.javabotblockapi","c":"Site","l":"TOP_GG"},{"p":"org.botblock.javabotblockapi.exceptions","c":"RatelimitedException","l":"toString()"},{"p":"org.botblock.javabotblockapi","c":"Site","l":"valueOf(String)","url":"valueOf(java.lang.String)"},{"p":"org.botblock.javabotblockapi.internal","c":"GetListAction.ApiField","l":"valueOf(String)","url":"valueOf(java.lang.String)"},{"p":"org.botblock.javabotblockapi","c":"Site","l":"values()"},{"p":"org.botblock.javabotblockapi.internal","c":"GetListAction.ApiField","l":"values()"},{"p":"org.botblock.javabotblockapi.annotations","c":"DeprecatedSince","l":"version()"},{"p":"org.botblock.javabotblockapi","c":"Site","l":"WONDERBOTLIST_COM"},{"p":"org.botblock.javabotblockapi","c":"Site","l":"YABL_XYZ"}]
