+++ docs/org/botblock/javabotblockapi/BotBlockAPI.html
@@ -2,10 +2,10 @@
 <!-- NewPage -->
 <html lang="de">
 <head>
-<!-- Generated by javadoc (11.0.4) on Fri Mar 06 18:26:13 CET 2020 -->
+<!-- Generated by javadoc (11.0.4) on Tue Mar 10 18:37:45 CET 2020 -->
 <title>BotBlockAPI (JavaBotBlockAPI 5.1.12 API)</title>
 <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
-<meta name="dc.created" content="2020-03-06">
+<meta name="dc.created" content="2020-03-10">
 <link rel="stylesheet" type="text/css" href="../../../stylesheet.css" title="Style">
 <link rel="stylesheet" type="text/css" href="../../../jquery/jquery-ui.css" title="Style">
 <script type="text/javascript" src="../../../script.js"></script>
@@ -135,7 +135,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <pre>public class <span class="typeNameLabel">BotBlockAPI</span>
 extends <a href="https://docs.oracle.com/javase/8/docs/api/java/lang/Object.html?is-external=true" title="class or interface in java.lang" class="externalLink">Object</a></pre>
 <div class="block">Class used to define the auth-tokens used for the different sites.
- <br>The instance of this class will be used in the <a href="requests/PostAction.html" title="class in org.botblock.javabotblockapi.requests"><code>PostAction</code></a> class.
+ <br>The instance of this class will be used in the <a href="internal/PostAction.html" title="class in org.botblock.javabotblockapi.internal"><code>PostAction</code></a> class.
  
  <p>Use the <a href="BotBlockAPI.Builder.html" title="class in org.botblock.javabotblockapi"><code>BotBlockAPI.Builder</code></a> class for easy creation.</div>
 </li>
