+++ docs/org/botblock/javabotblockapi/Site.html
@@ -2,10 +2,10 @@
 <!-- NewPage -->
 <html lang="de">
 <head>
-<!-- Generated by javadoc (11.0.4) on Fri Mar 06 18:26:13 CET 2020 -->
+<!-- Generated by javadoc (11.0.4) on Tue Mar 10 18:37:45 CET 2020 -->
 <title>Site (JavaBotBlockAPI 5.1.12 API)</title>
 <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
-<meta name="dc.created" content="2020-03-06">
+<meta name="dc.created" content="2020-03-10">
 <link rel="stylesheet" type="text/css" href="../../../stylesheet.css" title="Style">
 <link rel="stylesheet" type="text/css" href="../../../jquery/jquery-ui.css" title="Style">
 <script type="text/javascript" src="../../../script.js"></script>
@@ -147,8 +147,8 @@ extends <a href="https://docs.oracle.com/javase/8/docs/api/java/lang/Enum.html?i
  
  <p>The fields can be used in:
  <ul>
-     <li><a href="requests/GetBotAction.html" title="class in org.botblock.javabotblockapi.requests"><code>GetBotAction</code></a></li>
-     <li><a href="requests/GetListAction.html" title="class in org.botblock.javabotblockapi.requests"><code>GetListAction</code></a> or</li>
+     <li><a href="internal/GetBotActionImpl.html" title="class in org.botblock.javabotblockapi.internal"><code>GetBotAction</code></a></li>
+     <li><a href="internal/GetListAction.html" title="class in org.botblock.javabotblockapi.internal"><code>GetListAction</code></a> or</li>
      <li><a href="BotBlockAPI.Builder.html" title="class in org.botblock.javabotblockapi"><code>BotBlockAPI.Builder</code></a></li>
  </ul></div>
 <dl>
@@ -515,8 +515,8 @@ public static final&nbsp;<a href="Site.html" title="enum in org.botblock.javabot
              <br>This option still works, but is no longer supported, nor recommended!</div>
 </div>
 <div class="block"><a href="https://discordbots.org" target="_blank">discordbots.org</a>
- <br><b>This field may not be used in <a href="requests/GetBotAction.html" title="class in org.botblock.javabotblockapi.requests"><code>GetBotAction</code></a> and
- <a href="requests/GetListAction.html" title="class in org.botblock.javabotblockapi.requests"><code>GetListAction</code></a>!</b></div>
+ <br><b>This field may not be used in <a href="internal/GetBotActionImpl.html" title="class in org.botblock.javabotblockapi.internal"><code>GetBotAction</code></a> and
+ <a href="internal/GetListAction.html" title="class in org.botblock.javabotblockapi.internal"><code>GetListAction</code></a>!</b></div>
 </li>
 </ul>
 <a id="DISCORDBOTLIST_COM">
