+++ docs/org/botblock/javabotblockapi/package-tree.html
@@ -2,10 +2,10 @@
 <!-- NewPage -->
 <html lang="de">
 <head>
-<!-- Generated by javadoc (11.0.4) on Fri Mar 06 18:26:13 CET 2020 -->
+<!-- Generated by javadoc (11.0.4) on Tue Mar 10 18:37:45 CET 2020 -->
 <title>org.botblock.javabotblockapi Class Hierarchy (JavaBotBlockAPI 5.1.12 API)</title>
 <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
-<meta name="dc.created" content="2020-03-06">
+<meta name="dc.created" content="2020-03-10">
 <link rel="stylesheet" type="text/css" href="../../../stylesheet.css" title="Style">
 <link rel="stylesheet" type="text/css" href="../../../jquery/jquery-ui.css" title="Style">
 <script type="text/javascript" src="../../../script.js"></script>
