+++ docs/org/botblock/javabotblockapi/requests/package-summary.html
@@ -2,10 +2,10 @@
 <!-- NewPage -->
 <html lang="de">
 <head>
-<!-- Generated by javadoc (11.0.4) on Fri Mar 06 18:26:13 CET 2020 -->
+<!-- Generated by javadoc (11.0.4) on Tue Mar 10 18:37:45 CET 2020 -->
 <title>org.botblock.javabotblockapi.requests (JavaBotBlockAPI 5.1.12 API)</title>
 <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
-<meta name="dc.created" content="2020-03-06">
+<meta name="dc.created" content="2020-03-10">
 <link rel="stylesheet" type="text/css" href="../../../../stylesheet.css" title="Style">
 <link rel="stylesheet" type="text/css" href="../../../../jquery/jquery-ui.css" title="Style">
 <script type="text/javascript" src="../../../../script.js"></script>
@@ -97,58 +97,19 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <h1 title="Package" class="title">Package&nbsp;org.botblock.javabotblockapi.requests</h1>
 </div>
 <div class="contentContainer">
-<section role="region"><a id="package.description">
-<!--   -->
-</a>
-<div class="block">Contains classes used for various GET and POST requests.
- 
- <p>Use <a href="GetBotAction.html" title="class in org.botblock.javabotblockapi.requests"><code>GetBotAction</code></a> for retrieving bot information,
- <a href="GetListAction.html" title="class in org.botblock.javabotblockapi.requests"><code>GetListAction</code></a> for List related actions and
- <a href="PostAction.html" title="class in org.botblock.javabotblockapi.requests"><code>PostAction</code></a> for POST requests.</div>
-</section>
 <ul class="blockList">
 <li class="blockList">
 <table class="typeSummary">
-<caption><span>Class Summary</span><span class="tabEnd">&nbsp;</span></caption>
+<caption><span>Interface Summary</span><span class="tabEnd">&nbsp;</span></caption>
 <tr>
-<th class="colFirst" scope="col">Class</th>
+<th class="colFirst" scope="col">Interface</th>
 <th class="colLast" scope="col">Description</th>
 </tr>
 <tbody>
 <tr class="altColor">
-<th class="colFirst" scope="row"><a href="GetBotAction.html" title="class in org.botblock.javabotblockapi.requests">GetBotAction</a></th>
+<th class="colFirst" scope="row"><a href="GetBotAction.html" title="interface in org.botblock.javabotblockapi.requests">GetBotAction</a></th>
 <td class="colLast">
-<div class="block">Class used to perform GET actions on the <a href="https://botblock.org/api/docs#bots" target="_blank"><code>/api/bots/:id</code> endpoint</a>.</div>
-</td>
-</tr>
-<tr class="rowColor">
-<th class="colFirst" scope="row"><a href="GetListAction.html" title="class in org.botblock.javabotblockapi.requests">GetListAction</a></th>
-<td class="colLast">
-<div class="block">Class used to perform GET actions on the <a href="https://botblock.org/api/docs#lists" target="_blank"><code>/api/lists</code></a>
- and <a href="https://botblock.org/api/docs#list" target="_blank"><code>/api/lists/:id</code></a> endpoints.</div>
-</td>
-</tr>
-<tr class="altColor">
-<th class="colFirst" scope="row"><a href="PostAction.html" title="class in org.botblock.javabotblockapi.requests">PostAction</a></th>
-<td class="colLast">
-<div class="block">Class used to perform POST requests towards the <a href="https://botblock.org/api/docs#count" target="_blank">/api/count</a> endpoint.</div>
-</td>
-</tr>
-</tbody>
-</table>
-</li>
-<li class="blockList">
-<table class="typeSummary">
-<caption><span>Enum Summary</span><span class="tabEnd">&nbsp;</span></caption>
-<tr>
-<th class="colFirst" scope="col">Enum</th>
-<th class="colLast" scope="col">Description</th>
-</tr>
-<tbody>
-<tr class="altColor">
-<th class="colFirst" scope="row"><a href="GetListAction.ApiField.html" title="enum in org.botblock.javabotblockapi.requests">GetListAction.ApiField</a></th>
-<td class="colLast">
-<div class="block">Enum containing the different API fields a bot list may have.</div>
+<div class="block">Interface used to perform various GET actions towards the <a href="https://botblock.org/api/docs#bots" target="_blank"><code>/api/bots/:id</code></a> endpoint.</div>
 </td>
 </tr>
 </tbody>
