+++ docs/org/botblock/javabotblockapi/requests/package-tree.html
@@ -2,10 +2,10 @@
 <!-- NewPage -->
 <html lang="de">
 <head>
-<!-- Generated by javadoc (11.0.4) on Fri Mar 06 18:26:13 CET 2020 -->
+<!-- Generated by javadoc (11.0.4) on Tue Mar 10 18:37:45 CET 2020 -->
 <title>org.botblock.javabotblockapi.requests Class Hierarchy (JavaBotBlockAPI 5.1.12 API)</title>
 <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
-<meta name="dc.created" content="2020-03-06">
+<meta name="dc.created" content="2020-03-10">
 <link rel="stylesheet" type="text/css" href="../../../../stylesheet.css" title="Style">
 <link rel="stylesheet" type="text/css" href="../../../../jquery/jquery-ui.css" title="Style">
 <script type="text/javascript" src="../../../../script.js"></script>
@@ -102,29 +102,9 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 </div>
 <div class="contentContainer">
 <section role="region">
-<h2 title="Class Hierarchy">Class Hierarchy</h2>
+<h2 title="Interface Hierarchy">Interface Hierarchy</h2>
 <ul>
-<li class="circle">java.lang.<a href="https://docs.oracle.com/javase/8/docs/api/java/lang/Object.html?is-external=true" title="class or interface in java.lang" class="externalLink"><span class="typeNameLink">Object</span></a>
-<ul>
-<li class="circle">org.botblock.javabotblockapi.requests.<a href="GetBotAction.html" title="class in org.botblock.javabotblockapi.requests"><span class="typeNameLink">GetBotAction</span></a></li>
-<li class="circle">org.botblock.javabotblockapi.requests.<a href="GetListAction.html" title="class in org.botblock.javabotblockapi.requests"><span class="typeNameLink">GetListAction</span></a></li>
-<li class="circle">org.botblock.javabotblockapi.requests.<a href="PostAction.html" title="class in org.botblock.javabotblockapi.requests"><span class="typeNameLink">PostAction</span></a></li>
-</ul>
-</li>
-</ul>
-</section>
-<section role="region">
-<h2 title="Enum Hierarchy">Enum Hierarchy</h2>
-<ul>
-<li class="circle">java.lang.<a href="https://docs.oracle.com/javase/8/docs/api/java/lang/Object.html?is-external=true" title="class or interface in java.lang" class="externalLink"><span class="typeNameLink">Object</span></a>
-<ul>
-<li class="circle">java.lang.<a href="https://docs.oracle.com/javase/8/docs/api/java/lang/Enum.html?is-external=true" title="class or interface in java.lang" class="externalLink"><span class="typeNameLink">Enum</span></a>&lt;E&gt; (implements java.lang.<a href="https://docs.oracle.com/javase/8/docs/api/java/lang/Comparable.html?is-external=true" title="class or interface in java.lang" class="externalLink">Comparable</a>&lt;T&gt;, java.io.<a href="https://docs.oracle.com/javase/8/docs/api/java/io/Serializable.html?is-external=true" title="class or interface in java.io" class="externalLink">Serializable</a>)
-<ul>
-<li class="circle">org.botblock.javabotblockapi.requests.<a href="GetListAction.ApiField.html" title="enum in org.botblock.javabotblockapi.requests"><span class="typeNameLink">GetListAction.ApiField</span></a></li>
-</ul>
-</li>
-</ul>
-</li>
+<li class="circle">org.botblock.javabotblockapi.requests.<a href="GetBotAction.html" title="interface in org.botblock.javabotblockapi.requests"><span class="typeNameLink">GetBotAction</span></a></li>
 </ul>
 </section>
 </div>
