+++ docs/overview-tree.html
@@ -2,10 +2,10 @@
 <!-- NewPage -->
 <html lang="de">
 <head>
-<!-- Generated by javadoc (11.0.4) on Fri Mar 06 18:26:13 CET 2020 -->
+<!-- Generated by javadoc (11.0.4) on Tue Mar 10 18:37:45 CET 2020 -->
 <title>Class Hierarchy (JavaBotBlockAPI 5.1.12 API)</title>
 <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
-<meta name="dc.created" content="2020-03-06">
+<meta name="dc.created" content="2020-03-10">
 <link rel="stylesheet" type="text/css" href="stylesheet.css" title="Style">
 <link rel="stylesheet" type="text/css" href="jquery/jquery-ui.css" title="Style">
 <script type="text/javascript" src="script.js"></script>
@@ -100,6 +100,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <li><a href="org/botblock/javabotblockapi/package-tree.html">org.botblock.javabotblockapi</a>, </li>
 <li><a href="org/botblock/javabotblockapi/annotations/package-tree.html">org.botblock.javabotblockapi.annotations</a>, </li>
 <li><a href="org/botblock/javabotblockapi/exceptions/package-tree.html">org.botblock.javabotblockapi.exceptions</a>, </li>
+<li><a href="org/botblock/javabotblockapi/internal/package-tree.html">org.botblock.javabotblockapi.internal</a>, </li>
 <li><a href="org/botblock/javabotblockapi/requests/package-tree.html">org.botblock.javabotblockapi.requests</a></li>
 </ul>
 </div>
@@ -111,9 +112,9 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <ul>
 <li class="circle">org.botblock.javabotblockapi.<a href="org/botblock/javabotblockapi/BotBlockAPI.html" title="class in org.botblock.javabotblockapi"><span class="typeNameLink">BotBlockAPI</span></a></li>
 <li class="circle">org.botblock.javabotblockapi.<a href="org/botblock/javabotblockapi/BotBlockAPI.Builder.html" title="class in org.botblock.javabotblockapi"><span class="typeNameLink">BotBlockAPI.Builder</span></a></li>
-<li class="circle">org.botblock.javabotblockapi.requests.<a href="org/botblock/javabotblockapi/requests/GetBotAction.html" title="class in org.botblock.javabotblockapi.requests"><span class="typeNameLink">GetBotAction</span></a></li>
-<li class="circle">org.botblock.javabotblockapi.requests.<a href="org/botblock/javabotblockapi/requests/GetListAction.html" title="class in org.botblock.javabotblockapi.requests"><span class="typeNameLink">GetListAction</span></a></li>
-<li class="circle">org.botblock.javabotblockapi.requests.<a href="org/botblock/javabotblockapi/requests/PostAction.html" title="class in org.botblock.javabotblockapi.requests"><span class="typeNameLink">PostAction</span></a></li>
+<li class="circle">org.botblock.javabotblockapi.internal.<a href="org/botblock/javabotblockapi/internal/GetBotActionImpl.html" title="class in org.botblock.javabotblockapi.internal"><span class="typeNameLink">GetBotActionImpl</span></a> (implements org.botblock.javabotblockapi.requests.<a href="org/botblock/javabotblockapi/requests/GetBotAction.html" title="interface in org.botblock.javabotblockapi.requests">GetBotAction</a>)</li>
+<li class="circle">org.botblock.javabotblockapi.internal.<a href="org/botblock/javabotblockapi/internal/GetListAction.html" title="class in org.botblock.javabotblockapi.internal"><span class="typeNameLink">GetListAction</span></a></li>
+<li class="circle">org.botblock.javabotblockapi.internal.<a href="org/botblock/javabotblockapi/internal/PostAction.html" title="class in org.botblock.javabotblockapi.internal"><span class="typeNameLink">PostAction</span></a></li>
 <li class="circle">java.lang.<a href="https://docs.oracle.com/javase/8/docs/api/java/lang/Throwable.html?is-external=true" title="class or interface in java.lang" class="externalLink"><span class="typeNameLink">Throwable</span></a> (implements java.io.<a href="https://docs.oracle.com/javase/8/docs/api/java/io/Serializable.html?is-external=true" title="class or interface in java.io" class="externalLink">Serializable</a>)
 <ul>
 <li class="circle">java.lang.<a href="https://docs.oracle.com/javase/8/docs/api/java/lang/Exception.html?is-external=true" title="class or interface in java.lang" class="externalLink"><span class="typeNameLink">Exception</span></a>
@@ -132,6 +133,12 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 </ul>
 </section>
 <section role="region">
+<h2 title="Interface Hierarchy">Interface Hierarchy</h2>
+<ul>
+<li class="circle">org.botblock.javabotblockapi.requests.<a href="org/botblock/javabotblockapi/requests/GetBotAction.html" title="interface in org.botblock.javabotblockapi.requests"><span class="typeNameLink">GetBotAction</span></a></li>
+</ul>
+</section>
+<section role="region">
 <h2 title="Annotation Type Hierarchy">Annotation Type Hierarchy</h2>
 <ul>
 <li class="circle">org.botblock.javabotblockapi.annotations.<a href="org/botblock/javabotblockapi/annotations/DeprecatedSince.html" title="annotation in org.botblock.javabotblockapi.annotations"><span class="typeNameLink">DeprecatedSince</span></a> (implements java.lang.annotation.<a href="https://docs.oracle.com/javase/8/docs/api/java/lang/annotation/Annotation.html?is-external=true" title="class or interface in java.lang.annotation" class="externalLink">Annotation</a>)</li>
@@ -144,7 +151,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <ul>
 <li class="circle">java.lang.<a href="https://docs.oracle.com/javase/8/docs/api/java/lang/Enum.html?is-external=true" title="class or interface in java.lang" class="externalLink"><span class="typeNameLink">Enum</span></a>&lt;E&gt; (implements java.lang.<a href="https://docs.oracle.com/javase/8/docs/api/java/lang/Comparable.html?is-external=true" title="class or interface in java.lang" class="externalLink">Comparable</a>&lt;T&gt;, java.io.<a href="https://docs.oracle.com/javase/8/docs/api/java/io/Serializable.html?is-external=true" title="class or interface in java.io" class="externalLink">Serializable</a>)
 <ul>
-<li class="circle">org.botblock.javabotblockapi.requests.<a href="org/botblock/javabotblockapi/requests/GetListAction.ApiField.html" title="enum in org.botblock.javabotblockapi.requests"><span class="typeNameLink">GetListAction.ApiField</span></a></li>
+<li class="circle">org.botblock.javabotblockapi.internal.<a href="org/botblock/javabotblockapi/internal/GetListAction.ApiField.html" title="enum in org.botblock.javabotblockapi.internal"><span class="typeNameLink">GetListAction.ApiField</span></a></li>
 <li class="circle">org.botblock.javabotblockapi.<a href="org/botblock/javabotblockapi/Site.html" title="enum in org.botblock.javabotblockapi"><span class="typeNameLink">Site</span></a></li>
 </ul>
 </li>
