+++ docs/type-search-index.js
@@ -1,1 +1,1 @@
-typeSearchIndex = [{"l":"All Classes","url":"allclasses-index.html"},{"p":"org.botblock.javabotblockapi.requests","l":"GetListAction.ApiField"},{"p":"org.botblock.javabotblockapi","l":"BotBlockAPI"},{"p":"org.botblock.javabotblockapi","l":"BotBlockAPI.Builder"},{"p":"org.botblock.javabotblockapi.annotations","l":"DeprecatedSince"},{"p":"org.botblock.javabotblockapi.requests","l":"GetBotAction"},{"p":"org.botblock.javabotblockapi.requests","l":"GetListAction"},{"p":"org.botblock.javabotblockapi.requests","l":"PostAction"},{"p":"org.botblock.javabotblockapi.exceptions","l":"RatelimitedException"},{"p":"org.botblock.javabotblockapi","l":"Site"}]
+typeSearchIndex = [{"l":"All Classes","url":"allclasses-index.html"},{"p":"org.botblock.javabotblockapi.internal","l":"GetListAction.ApiField"},{"p":"org.botblock.javabotblockapi","l":"BotBlockAPI"},{"p":"org.botblock.javabotblockapi","l":"BotBlockAPI.Builder"},{"p":"org.botblock.javabotblockapi.annotations","l":"DeprecatedSince"},{"p":"org.botblock.javabotblockapi.requests","l":"GetBotAction"},{"p":"org.botblock.javabotblockapi.internal","l":"GetBotActionImpl"},{"p":"org.botblock.javabotblockapi.internal","l":"GetListAction"},{"p":"org.botblock.javabotblockapi.internal","l":"PostAction"},{"p":"org.botblock.javabotblockapi.exceptions","l":"RatelimitedException"},{"p":"org.botblock.javabotblockapi","l":"Site"}]
