+++ src/main/java/org/botblock/javabotblockapi/internal/CheckUtil.java
@@ -16,11 +16,13 @@
  * OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-/**
- * Contains classes used for various GET and POST requests.
- * 
- * <p>Use {@link org.botblock.javabotblockapi.requests.GetBotAction GetBotAction} for retrieving bot information,
- * {@link org.botblock.javabotblockapi.requests.GetListAction GetListAction} for List related actions and
- * {@link org.botblock.javabotblockapi.requests.PostAction PostAction} for POST requests.
- */
-package org.botblock.javabotblockapi.requests;
+package org.botblock.javabotblockapi.internal;
+
+class CheckUtil{
+    
+    public static void notEmpty(String value, String name){
+        if(value.isEmpty())
+            throw new NullPointerException(name + " may not be empty.");
+    }
+    
+}
